Convert test_velocity_integrals.jl to a CI test
